Simplify Hurkle (C#) folder structure
