+++ 51_Hurkle/csharp/CardinalDirection.cs
