+++ 51_Hurkle/csharp/ConsoleHurkleView.cs
