+++ 51_Hurkle/csharp/FailedGuessViewModel.cs
