+++ 51_Hurkle/csharp/Hurkle.sln
@@ -3,25 +3,23 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.30114.105
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{EBE7BC2B-8F2E-41D5-AF36-7AAC7CE0E1BF}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "hurkle", "src\hurkle\hurkle.csproj", "{47578EC1-A012-4BF7-8709-64F675E72DB0}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Hurkle", "Hurkle.csproj", "{BE321D5B-93BD-4F91-A875-564DC9D4094F}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{BE321D5B-93BD-4F91-A875-564DC9D4094F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BE321D5B-93BD-4F91-A875-564DC9D4094F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BE321D5B-93BD-4F91-A875-564DC9D4094F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BE321D5B-93BD-4F91-A875-564DC9D4094F}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{47578EC1-A012-4BF7-8709-64F675E72DB0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{47578EC1-A012-4BF7-8709-64F675E72DB0}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{47578EC1-A012-4BF7-8709-64F675E72DB0}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{47578EC1-A012-4BF7-8709-64F675E72DB0}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(NestedProjects) = preSolution
-		{47578EC1-A012-4BF7-8709-64F675E72DB0} = {EBE7BC2B-8F2E-41D5-AF36-7AAC7CE0E1BF}
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {42DC6AE5-5127-4B1B-BD5E-F3B1CCDC3822}
 	EndGlobalSection
 EndGlobal
