+++ 51_Hurkle/csharp/IHurkleView.cs
