+++ 51_Hurkle/csharp/LossViewModel.cs
